Refactor plugins into separate files. Signed-off-by: William Juan <66322422+wjuan-AFK@users.noreply.github.com>
